+++ firmware/include/led_control.h
@@ -9,13 +9,13 @@ enum LampState { OFF, ON };
 
 extern LampState currentLampState;
 extern uint8_t currentMode;           // MODE_WARM or MODE_COOL
-extern uint8_t brightness;            // 0 to MAX_BRIGHTNESS
+extern uint16_t brightness;           // 0 to MAX_BRIGHTNESS (12-bit — see config.h)
 extern int8_t brightnessDirection;    // -1 or +1
 
 void initLED();
 
 // Power control
-void turnOn(uint8_t mode, uint8_t brightnessLevel);
+void turnOn(uint8_t mode, uint16_t brightnessLevel);
 void turnOff();
 
 // Turns on at brightness 0 with no crossfade — for callers that immediately follow with
@@ -26,12 +26,12 @@ void turnOff();
 void turnOnAtZero(uint8_t mode);
 
 // Mode swap with crossfade
-void swapMode(uint8_t newMode, uint8_t newBrightness);
+void swapMode(uint8_t newMode, uint16_t newBrightness);
 
 // Brightness
 void incrementBrightness();
 void reverseBrightnessDirection();
-uint8_t getActiveBrightness();
+uint16_t getActiveBrightness();
 
 // Smooth transition (call from loop)
 void updateModeTransition();
@@ -40,7 +40,7 @@ void updateModeTransition();
 // desired brightness; updateBrightnessSlew() (call every loop) eases the live brightness
 // and PWM output toward that target with exponential smoothing, so fast pot movements
 // produce a smooth, slightly lagging ramp instead of an instant jump.
-void setBrightnessTarget(uint8_t target);
+void setBrightnessTarget(uint16_t target);
 void updateBrightnessSlew();
 
 // Battery indicator pulse (non-blocking)
